+++ src/utils/index.ts
@@ -201,6 +201,47 @@ export async function fetchResource(resourceType: string, resourceId: string, ha
     }
 }
 
+export async function searchResource(
+    resourceType: string,
+    query: string,
+    limit?: number,
+    skip?: number,
+    hazardType?: string
+) {
+    let url = config.hostname;
+
+    if (resourceType.toLowerCase() === "dataset") {
+        url = `${url}/data/api/datasets/search?text=${query}&limit=${limit ?? 5}&skip=${skip ?? 0}`;
+    } else if (resourceType.toLowerCase() === "hazard") {
+        url = `${url}/hazard/api/${hazardType}/search?text=${query}&limit=${limit ?? 5}&skip=${skip ?? 0}`;
+    } else if (resourceType.toLowerCase().includes("mapping")) {
+        url = `${url}/dfr3/api/mappings/search?text=${query}&limit=${limit ?? 5}&skip=${skip ?? 0}`;
+    }
+
+    try {
+        const response = await axios.get(url, { headers: getHeaders() });
+
+        // Check if the response has data
+        if (!response.data || Object.keys(response.data).length === 0) {
+            return { error: `No data found for resourceType: ${resourceType}, query: ${query}` };
+        }
+
+        return response.data;
+    } catch (error: any) {
+        // Handle axios-specific errors
+        if (axios.isAxiosError(error)) {
+            return {
+                error: error.response
+                    ? `Error: ${error.response.status} - ${error.response.data}`
+                    : `Network or unknown error: ${error.message}`
+            };
+        }
+
+        // Handle unexpected errors
+        return { error: `Unexpected error: ${error.message}` };
+    }
+}
+
 export function toSingular(disaster: string): string {
     // Mapping of plural to singular forms
     const singularMapping: { [key: string]: string } = {
@@ -256,6 +297,75 @@ export function getStatusColor(status?: string) {
     }
 }
 
+export const getOutputDatasetIDsFromExecutionFile = (
+    execution: DatawolfExecutionFile,
+    workflow: DatawolfWorkflowFile
+): string[] => {
+    const outputDatasetIDs: string[] = [];
+    if (!execution || !workflow || !workflow.steps) {
+        return outputDatasetIDs;
+    }
+
+    workflow.steps.forEach((step) => {
+        Object.entries(step.outputs).forEach(([outputKey, outputValue]) => {
+            if (
+                outputValue !== "" &&
+                outputValue !== null &&
+                execution.datasets[outputValue] !== "" &&
+                execution.datasets[outputValue] !== null &&
+                execution.datasets[outputValue] !== "ERROR" &&
+                outputKey !== "stdout"
+            ) {
+                outputDatasetIDs.push(execution.datasets[outputValue]);
+            }
+        });
+    });
+
+    return outputDatasetIDs;
+};
+
+export const getOutputDatasetIDsFromWorkflows = async (workflows: DatawolfWorkflowFile[]): Promise<string[]> => {
+    const outputDatasetIDs: string[] = [];
+    for (const workflow of workflows) {
+        try {
+            const executions = await axios.get(`${config.datawolfApi}/workflows/${workflow.id ?? ""}/executions`, {
+                headers: getHeaders()
+            });
+            if (executions.data) {
+                for (const execution of executions.data) {
+                    outputDatasetIDs.push(...getOutputDatasetIDsFromExecutionFile(execution, workflow));
+                }
+            }
+        } catch (error: any) {
+            // Handle axios-specific errors
+            if (axios.isAxiosError(error)) {
+                console.error(
+                    "Error fetching executions:",
+                    error.response
+                        ? `Error: ${error.response.status} - ${error.response.data}`
+                        : `Network or unknown error: ${error.message}`
+                );
+            }
+
+            // Handle unexpected errors
+            console.error("Unexpected error:", error.message);
+        }
+    }
+    return outputDatasetIDs;
+};
+
+export function downloadMetadata(data: any): void {
+    const metadata = JSON.stringify(data, null, 2); // Example metadata formatting
+    const blob = new Blob([metadata], { type: "application/json" });
+    const url = URL.createObjectURL(blob);
+
+    const a = document.createElement("a");
+    a.href = url;
+    a.download = `${data.name || data.title || "metadata"}.json`;
+    a.click();
+    URL.revokeObjectURL(url);
+}
+
 // export function attachExecutionStatusToSteps(execution: DatawolfExecutionFile, workflow: Workflow) {
 //     if (!execution || !workflow || !workflow.steps) {
 //         throw new Error("Invalid execution or workflow data.");
@@ -273,3 +383,13 @@ export function getStatusColor(status?: string) {
 //         };
 //     });
 // }
+
+export function breakCamelCaseAndCapitalize(camelCaseString: string): string {
+    return camelCaseString
+        .replace(/([a-z])([A-Z])/g, "$1 $2") // Add space before uppercase letters
+        .replace(/([A-Z])([A-Z][a-z])/g, "$1 $2") // Handle consecutive uppercase letters
+        .trim() // Remove leading/trailing spaces
+        .split(" ") // Split into words
+        .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase()) // Capitalize each word
+        .join(" "); // Join the words back into a phrase
+}